+ enh. add static Lexer function reference to serve generic functionallity (1st case: track style changes)

+++ src/StyleLexers/EditLexer.h
@@ -20,12 +20,12 @@
 
 typedef enum
 {
-  FCT_SETTING_CHANGE, // value -1: query changed, 0: set->not changed, 1: set->has changed
+  FCT_SETTING_CHANGE, // value -b: rest style bit, 0: get bit-set, +b: set style bit
   FCT_PASS
 }
 LexFunctionType;
 
-typedef int (*LexFunctionPtr_t)(LexFunctionType type, int value);
+typedef __int64 (*LexFunctionPtr_t)(LexFunctionType type, int value);
 
 
 // -----------------------------------------------------------------------------
